Perkembangan Usaha Mikro Kecil dan Menengah (UMKM)

Dataset ini berisi mengenai kondisi dan perkembangan UMKM di Indonesia. Catatan: 1. Sumbangan PDB UMKM pada tahun 1997-2000 masih didasarkan atas harga berlaku. Sumbangan PDB UMKM pada tahun 2001-2012 didasarkan atas harga konstan tahun 2000; 2. Sumbangan PDB mencakup migas dan non migas; 3. Perhitungan sumbangan PDB UMKM pada tahun 2012 masih sangat sementara dan saat ini masih dihitung ulang; 4. Nilai Ekspor UMKM hanya didasarkan pada data di sektor pertanian (pertanian, perikanan, kelautan, peternakan, kehutanan, perkebunan), industri pengolahan, dan pertambangan/penggalian

ids_umkm

Format

Dataframe dengan 16 baris dan 9 kolom

tahun

Tahun

jumlah

Jumlah UMKM

pertumbuhan_jumla

Pertumbuhan jumlah UMKM (persen)

jumlah_tenaga_kerja

Jumlah tenaga kerja UMKM (orang)

pertumbuhan_jumlah_tenaga_kerja

Pertumbuhan jumlah tenaga kerja UMKM (persen)

sumbangan_pdb

Sumbangan PDB UMKM (harga konstan, Rp. Miliar)

sumbangan_pdb

Pertumbuhan sumbangan PDB UMKM (persen)

nilai_ekspor

Nilai ekspor UMKM (Rp. Miliar)

pertumbuhan_nilai_ekspor

Pertumbuhan nilai ekspor UMKM (persen)

Source

Portal Satu Data Indonesia https://data.go.id/dataset/perkembangan-usaha-mikro-kecil-dan-menengah-umkm